Shuffle permission decoding logic
