+++ packages/7715-permission-types/src/permissions/caveats/nativeTokenAllowance.ts
@@ -1,28 +1,29 @@
-import { hexToBigInt, hexToNumber } from '@metamask/utils';
+import { hexToNumber } from '@metamask/utils';
 
-import { expiryRule } from './expiryRule';
-import { nativePayeeRuleDecoder } from './nativePayeeRuleDecoder';
-import { redeemerRuleDecoder } from './redeemerRuleDecoder';
+import { expiryRule } from '../rules/expiry';
+import { nativePayeeRuleDecoder } from '../rules/payee';
+import { redeemerRuleDecoder } from '../rules/redeemer';
 import type {
   ChecksumCaveat,
   ChecksumEnforcersByChainId,
   DecodedPermission,
   MakePermissionDecoderConfig,
-} from './types';
+} from '../types';
 import {
   getByteLength,
   getTermsByEnforcer,
-  MAX_PERIOD_DURATION,
   splitHex,
-} from './utils';
+  UINT256_MAX,
+  ZERO_32_BYTES,
+} from '../utils';
 
 /**
- * Builds the configuration for the native-token-periodic permission decoder.
+ * Builds the configuration for the native-token-allowance permission decoder.
  *
  * @param contractAddresses - Checksummed enforcer addresses for the chain.
- * @returns The native-token-periodic permission decoder configuration.
+ * @returns The native-token-allowance permission decoder configuration.
  */
-export function makeNativeTokenPeriodicDecoderConfig(
+export function makeNativeTokenAllowanceDecoderConfig(
   contractAddresses: ChecksumEnforcersByChainId,
 ): MakePermissionDecoderConfig {
   const {
@@ -35,7 +36,7 @@ export function makeNativeTokenPeriodicDecoderConfig(
   } = contractAddresses;
 
   return {
-    permissionType: 'native-token-periodic',
+    permissionType: 'native-token-allowance',
     contractAddresses,
     optionalEnforcers: [
       timestampEnforcer, // expiry rule
@@ -53,11 +54,11 @@ export function makeNativeTokenPeriodicDecoderConfig(
 }
 
 /**
- * Decodes native-token-periodic permission data from caveats; throws on invalid.
+ * Decodes native-token-allowance permission data from caveats; throws on invalid.
  *
  * @param caveats - Caveats from the permission context (checksummed).
  * @param contractAddresses - Checksummed enforcer addresses for the chain.
- * @returns Decoded periodic terms.
+ * @returns Decoded allowance terms.
  */
 function validateAndDecodeData(
   caveats: ChecksumCaveat[],
@@ -83,44 +84,36 @@ function validateAndDecodeData(
   const EXPECTED_TERMS_BYTELENGTH = 96; // 32 + 32 + 32
 
   if (getByteLength(terms) !== EXPECTED_TERMS_BYTELENGTH) {
-    throw new Error('Invalid native-token-periodic terms: expected 96 bytes');
+    throw new Error('Invalid native-token-allowance terms: expected 96 bytes');
   }
 
-  const [periodAmount, periodDurationRaw, startTimeRaw] = splitHex(
+  const [allowanceAmount, periodDurationRaw, startTimeRaw] = splitHex(
     terms,
     [32, 32, 32],
   );
-  if (!periodAmount || !periodDurationRaw || !startTimeRaw) {
-    throw new Error('Invalid native-token-periodic terms');
+  if (!allowanceAmount || !periodDurationRaw || !startTimeRaw) {
+    throw new Error('Invalid native-token-allowance terms');
   }
 
-  const periodDuration = hexToNumber(periodDurationRaw);
+  if (periodDurationRaw.toLowerCase() !== UINT256_MAX) {
+    throw new Error(
+      'Invalid native-token-allowance terms: periodDuration must be UINT256_MAX',
+    );
+  }
+
   const startTime = hexToNumber(startTimeRaw);
-  const periodAmountBigInt = hexToBigInt(periodAmount);
-
-  if (periodAmountBigInt === 0n) {
-    throw new Error(
-      'Invalid native-token-periodic terms: periodAmount must be a positive number',
-    );
-  }
-
-  if (periodDuration === 0) {
-    throw new Error(
-      'Invalid native-token-periodic terms: periodDuration must be a positive number',
-    );
-  }
-
-  if (periodDuration > MAX_PERIOD_DURATION) {
-    throw new Error(
-      'Invalid native-token-periodic terms: periodDuration must be less than or equal to MAX_PERIOD_DURATION',
-    );
-  }
 
   if (startTime === 0) {
     throw new Error(
-      'Invalid native-token-periodic terms: startTime must be a positive number',
+      'Invalid native-token-allowance terms: startTime must be a positive number',
     );
   }
 
-  return { periodAmount, periodDuration, startTime };
+  if (allowanceAmount === ZERO_32_BYTES) {
+    throw new Error(
+      'Invalid native-token-allowance terms: allowanceAmount must be a positive number',
+    );
+  }
+
+  return { allowanceAmount, startTime };
 }
