--- packages/7715-permission-types/src/permissions/payeeRule.ts
@@ -1,12 +1,0 @@
-import type { Hex } from '../types';
-
-/**
- * Execution permission rule restricting which addresses may receive payments
- * (on-chain AllowedCalldataEnforcer / AllowedTargetsEnforcer caveat).
- */
-export type PayeeRule = {
-  type: 'payee';
-  data: {
-    addresses: Hex[];
-  };
-};
