+++ packages/7715-permission-types/test/permissions/caveats/erc20TokenPeriodic.test.ts
@@ -8,8 +8,8 @@ import {
   DELEGATOR_CONTRACTS,
 } from '@metamask/delegation-deployments';
 
-import { createPermissionDecodersForContracts } from '../src/permissions';
-import { MAX_PERIOD_DURATION, ZERO_32_BYTES } from '../src/permissions/utils';
+import { createPermissionDecodersForContracts } from '../../../src/permissions';
+import { MAX_PERIOD_DURATION, ZERO_32_BYTES } from '../../../src/permissions/utils';
 
 describe('erc20-token-periodic decoder', () => {
   const chainId = CHAIN_ID.sepolia;
