+++ packages/7715-permission-types/test/permissions/caveats/erc20TokenRevocation.test.ts
@@ -5,7 +5,7 @@ import {
 } from '@metamask/delegation-deployments';
 import { getChecksumAddress } from '@metamask/utils';
 
-import { createPermissionDecodersForContracts } from '../src/permissions';
+import { createPermissionDecodersForContracts } from '../../../src/permissions';
 
 describe('erc20-token-revocation decoder', () => {
   const chainId = CHAIN_ID.sepolia;
