+++ packages/7715-permission-types/test/permissions/caveats/nativeTokenAllowance.test.ts
@@ -5,17 +5,19 @@ import {
   DELEGATOR_CONTRACTS,
 } from '@metamask/delegation-deployments';
 
-import { createPermissionDecodersForContracts } from '../src/permissions';
-import { ZERO_32_BYTES } from '../src/permissions/utils';
+import { createPermissionDecodersForContracts } from '../../../src/permissions';
 
-describe('erc20-token-allowance decoder', () => {
+describe('native-token-allowance decoder', () => {
   const chainId = CHAIN_ID.sepolia;
   const contracts = DELEGATOR_CONTRACTS['1.3.0'][chainId];
-  const { TimestampEnforcer, ERC20PeriodTransferEnforcer, ValueLteEnforcer } =
-    contracts;
+  const {
+    TimestampEnforcer,
+    NativeTokenPeriodTransferEnforcer,
+    ExactCalldataEnforcer,
+  } = contracts;
   const permissionDecoders = createPermissionDecodersForContracts(contracts);
   const decoder = permissionDecoders.find(
-    (candidate) => candidate.permissionType === 'erc20-token-allowance',
+    (candidate) => candidate.permissionType === 'native-token-allowance',
   );
   if (!decoder) {
     throw new Error('Decoder not found');
@@ -30,31 +32,29 @@ describe('erc20-token-allowance decoder', () => {
     args: '0x' as const,
   };
 
-  const valueLteCaveat = {
-    enforcer: ValueLteEnforcer,
-    terms: ZERO_32_BYTES,
+  const exactCalldataCaveat = {
+    enforcer: ExactCalldataEnforcer,
+    terms: '0x' as Hex,
     args: '0x' as const,
   };
 
-  const TOKEN_ADDRESS_HEX = 'aa'.repeat(20);
-  const TOKEN_ADDRESS: Hex = `0x${TOKEN_ADDRESS_HEX}`;
-  const ALLOWANCE_AMOUNT_HEX = 100n.toString(16).padStart(64, '0');
+  const PERIOD_AMOUNT_HEX = 100n.toString(16).padStart(64, '0');
   const PERIOD_DURATION_MAX_HEX = 'f'.repeat(64);
   const START_DATE_HEX = (1715664).toString(16).padStart(64, '0');
   const ALLOWANCE_TERMS =
-    `0x${TOKEN_ADDRESS_HEX}${ALLOWANCE_AMOUNT_HEX}${PERIOD_DURATION_MAX_HEX}${START_DATE_HEX}` as Hex;
+    `0x${PERIOD_AMOUNT_HEX}${PERIOD_DURATION_MAX_HEX}${START_DATE_HEX}` as Hex;
 
-  it('rejects duplicate ERC20PeriodTransferEnforcer caveats', () => {
+  it('rejects duplicate NativeTokenPeriodTransferEnforcer caveats', () => {
     const caveats = [
       expiryCaveat,
-      valueLteCaveat,
+      exactCalldataCaveat,
       {
-        enforcer: ERC20PeriodTransferEnforcer,
+        enforcer: NativeTokenPeriodTransferEnforcer,
         terms: ALLOWANCE_TERMS,
         args: '0x' as const,
       },
       {
-        enforcer: ERC20PeriodTransferEnforcer,
+        enforcer: NativeTokenPeriodTransferEnforcer,
         terms: ALLOWANCE_TERMS,
         args: '0x' as const,
       },
@@ -62,6 +62,7 @@ describe('erc20-token-allowance decoder', () => {
     const result = decoder.validateAndDecodePermission(caveats);
     expect(result.isValid).toBe(false);
 
+    // this is here as a type guard
     if (result.isValid) {
       throw new Error('Expected invalid result');
     }
@@ -70,13 +71,13 @@ describe('erc20-token-allowance decoder', () => {
   });
 
   it('rejects truncated terms', () => {
-    const truncatedTerms: Hex = `0x${'00'.repeat(40)}`; // 40 bytes, need 116
+    const truncatedTerms: Hex = `0x${'00'.repeat(40)}`; // 40 bytes, need 96
 
     const caveats = [
       expiryCaveat,
-      valueLteCaveat,
+      exactCalldataCaveat,
       {
-        enforcer: ERC20PeriodTransferEnforcer,
+        enforcer: NativeTokenPeriodTransferEnforcer,
         terms: truncatedTerms,
         args: '0x' as const,
       },
@@ -90,7 +91,7 @@ describe('erc20-token-allowance decoder', () => {
     }
 
     expect(result.error.message).toContain(
-      'Invalid erc20-token-allowance terms: expected 116 bytes',
+      'Invalid native-token-allowance terms: expected 96 bytes',
     );
   });
 
@@ -98,9 +99,9 @@ describe('erc20-token-allowance decoder', () => {
     const termsWithTrailing = `${ALLOWANCE_TERMS}deadbeef` as Hex;
     const caveats = [
       expiryCaveat,
-      valueLteCaveat,
+      exactCalldataCaveat,
       {
-        enforcer: ERC20PeriodTransferEnforcer,
+        enforcer: NativeTokenPeriodTransferEnforcer,
         terms: termsWithTrailing,
         args: '0x' as const,
       },
@@ -113,21 +114,20 @@ describe('erc20-token-allowance decoder', () => {
     }
 
     expect(result.error.message).toContain(
-      'Invalid erc20-token-allowance terms: expected 116 bytes',
+      'Invalid native-token-allowance terms: expected 96 bytes',
     );
   });
 
-  it('rejects when ValueLteEnforcer terms are not zero', () => {
-    const nonZeroValueLte: Hex = `0x${'0'.repeat(62)}01` as Hex;
+  it('rejects when ExactCalldataEnforcer terms are not 0x', () => {
     const caveats = [
       expiryCaveat,
       {
-        enforcer: ValueLteEnforcer,
-        terms: nonZeroValueLte,
+        enforcer: ExactCalldataEnforcer,
+        terms: '0x00' as Hex,
         args: '0x' as const,
       },
       {
-        enforcer: ERC20PeriodTransferEnforcer,
+        enforcer: NativeTokenPeriodTransferEnforcer,
         terms: ALLOWANCE_TERMS,
         args: '0x' as const,
       },
@@ -140,16 +140,16 @@ describe('erc20-token-allowance decoder', () => {
     }
 
     expect(result.error.message).toContain(
-      `Invalid value-lte terms: must be ${ZERO_32_BYTES}`,
+      'Invalid exact-calldata terms: must be 0x',
     );
   });
 
-  it('successfully decodes valid erc20-token-allowance caveats', () => {
+  it('successfully decodes valid native-token-allowance caveats', () => {
     const caveats = [
       expiryCaveat,
-      valueLteCaveat,
+      exactCalldataCaveat,
       {
-        enforcer: ERC20PeriodTransferEnforcer,
+        enforcer: NativeTokenPeriodTransferEnforcer,
         terms: ALLOWANCE_TERMS,
         args: '0x' as const,
       },
@@ -164,8 +164,7 @@ describe('erc20-token-allowance decoder', () => {
 
     expect(result.expiry).toBe(1720000);
     expect(result.data).toStrictEqual({
-      tokenAddress: TOKEN_ADDRESS,
-      allowanceAmount: `0x${ALLOWANCE_AMOUNT_HEX}`,
+      allowanceAmount: `0x${PERIOD_AMOUNT_HEX}`,
       startTime: 1715664,
     });
   });
@@ -173,13 +172,13 @@ describe('erc20-token-allowance decoder', () => {
   it('successfully decodes when periodDuration uses uppercase UINT256_MAX', () => {
     const uppercaseMax = 'F'.repeat(64);
     const terms =
-      `0x${TOKEN_ADDRESS_HEX}${ALLOWANCE_AMOUNT_HEX}${uppercaseMax}${START_DATE_HEX}` as Hex;
+      `0x${PERIOD_AMOUNT_HEX}${uppercaseMax}${START_DATE_HEX}` as Hex;
 
     const caveats = [
       expiryCaveat,
-      valueLteCaveat,
+      exactCalldataCaveat,
       {
-        enforcer: ERC20PeriodTransferEnforcer,
+        enforcer: NativeTokenPeriodTransferEnforcer,
         terms,
         args: '0x' as const,
       },
@@ -192,13 +191,13 @@ describe('erc20-token-allowance decoder', () => {
   it('rejects when periodDuration is not UINT256_MAX', () => {
     const nonMaxDuration = (86400).toString(16).padStart(64, '0');
     const terms =
-      `0x${TOKEN_ADDRESS_HEX}${ALLOWANCE_AMOUNT_HEX}${nonMaxDuration}${START_DATE_HEX}` as Hex;
+      `0x${PERIOD_AMOUNT_HEX}${nonMaxDuration}${START_DATE_HEX}` as Hex;
 
     const caveats = [
       expiryCaveat,
-      valueLteCaveat,
+      exactCalldataCaveat,
       {
-        enforcer: ERC20PeriodTransferEnforcer,
+        enforcer: NativeTokenPeriodTransferEnforcer,
         terms,
         args: '0x' as const,
       },
@@ -211,19 +210,19 @@ describe('erc20-token-allowance decoder', () => {
     }
 
     expect(result.error.message).toContain(
-      'Invalid erc20-token-allowance terms: periodDuration must be UINT256_MAX',
+      'Invalid native-token-allowance terms: periodDuration must be UINT256_MAX',
     );
   });
 
   it('rejects when startTime is 0', () => {
     const startTimeZero = '0'.repeat(64);
     const terms =
-      `0x${TOKEN_ADDRESS_HEX}${ALLOWANCE_AMOUNT_HEX}${PERIOD_DURATION_MAX_HEX}${startTimeZero}` as Hex;
+      `0x${PERIOD_AMOUNT_HEX}${PERIOD_DURATION_MAX_HEX}${startTimeZero}` as Hex;
     const caveats = [
       expiryCaveat,
-      valueLteCaveat,
+      exactCalldataCaveat,
       {
-        enforcer: ERC20PeriodTransferEnforcer,
+        enforcer: NativeTokenPeriodTransferEnforcer,
         terms,
         args: '0x' as const,
       },
@@ -236,20 +235,20 @@ describe('erc20-token-allowance decoder', () => {
     }
 
     expect(result.error.message).toContain(
-      'Invalid erc20-token-allowance terms: startTime must be a positive number',
+      'Invalid native-token-allowance terms: startTime must be a positive number',
     );
   });
 
   it('rejects when allowanceAmount is 0', () => {
     const allowanceZero = '0'.repeat(64);
     const terms =
-      `0x${TOKEN_ADDRESS_HEX}${allowanceZero}${PERIOD_DURATION_MAX_HEX}${START_DATE_HEX}` as Hex;
+      `0x${allowanceZero}${PERIOD_DURATION_MAX_HEX}${START_DATE_HEX}` as Hex;
 
     const caveats = [
       expiryCaveat,
-      valueLteCaveat,
+      exactCalldataCaveat,
       {
-        enforcer: ERC20PeriodTransferEnforcer,
+        enforcer: NativeTokenPeriodTransferEnforcer,
         terms,
         args: '0x' as const,
       },
@@ -263,7 +262,7 @@ describe('erc20-token-allowance decoder', () => {
     }
 
     expect(result.error.message).toContain(
-      'Invalid erc20-token-allowance terms: allowanceAmount must be a positive number',
+      'Invalid native-token-allowance terms: allowanceAmount must be a positive number',
     );
   });
 });
